Adds `ColumnHeader` component to `Table` (#2178)
