+++ .changeset/lucky-goats-kneel.md
@@ -1,0 +1,5 @@
+---
+'@itwin/itwinui-react': patch
+---
+
+Reduced layout thrashing on `Table` component by memoizing an expensive `ref` function.
