+++ packages/itwinui-react/src/core/Table/Table.tsx
@@ -30,26 +30,17 @@ import { ProgressRadial } from '../ProgressIndicators/ProgressRadial.js';
 import {
   useGlobals,
   useResizeObserver,
-  SvgSortDown,
-  SvgSortUp,
   useLayoutEffect,
   Box,
   createWarningLogger,
   ShadowRoot,
-  LineClamp,
   useMergedRefs,
   useLatestRef,
   useVirtualScroll,
 } from '../../utils/index.js';
 import type { CommonProps } from '../../utils/index.js';
-import {
-  TableColumnsContext,
-  getCellStyle,
-  getStickyStyle,
-  getSubRowStyle,
-} from './utils.js';
+import { TableColumnsContext } from './utils.js';
 import { TableRowMemoized } from './TableRowMemoized.js';
-import { FilterToggle } from './filters/index.js';
 import type { TableFilterValue } from './filters/index.js';
 import { customFilterFunctions } from './filters/customFilterFunctions.js';
 import {
@@ -73,16 +64,12 @@ import {
 } from './actionHandlers/index.js';
 import { SELECTION_CELL_ID } from './columns/index.js';
 import { Virtualizer, type VirtualItem } from '@tanstack/react-virtual';
+import { ColumnHeader } from './ColumnHeader.js';
 
 const singleRowSelectedAction = 'singleRowSelected';
 const shiftRowSelectedAction = 'shiftRowSelected';
 export const tableResizeStartAction = 'tableResizeStart';
 const tableResizeEndAction = 'tableResizeEnd';
-
-const COLUMN_MIN_WIDTHS = {
-  default: 72,
-  withExpander: 108, // expander column should be wider to accommodate the expander icon
-};
 
 const logWarning = createWarningLogger();
 
@@ -668,12 +655,6 @@ export const Table = <
       (column) => column.filterValue != null && column.filterValue !== '',
     ) || !!globalFilterValue;
 
-  const showFilterButton = (column: HeaderGroup<T>) =>
-    (data.length !== 0 || areFiltersSet) && column.canFilter && !!column.Filter;
-
-  const showSortButton = (column: HeaderGroup<T>) =>
-    data.length !== 0 && column.canSort;
-
   const onRowClickHandler = React.useCallback(
     (event: React.MouseEvent, row: Row<T>) => {
       const isDisabled = isRowDisabled?.(row.original);
@@ -924,18 +905,6 @@ export const Table = <
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const isHeaderDirectClick = React.useRef(false);
-
-  const columnResizeRef = React.useCallback(
-    (el: HTMLDivElement | null, column: HeaderGroup<T>) => {
-      if (el) {
-        columnRefs.current[column.id] = el;
-        column.resizeWidth = el.getBoundingClientRect().width;
-      }
-    },
-    [],
-  );
-
   return (
     <TableColumnsContext.Provider
       value={columns as Column<Record<string, unknown>>[]}
@@ -987,155 +956,25 @@ export const Table = <
               >
                 <Box {...headerGroupProps}>
                   {headerGroup.headers.map((column, index) => {
-                    const { onClick, ...restSortProps } =
-                      column.getSortByToggleProps();
-
-                    const columnHasExpanders =
-                      hasAnySubRows &&
-                      index ===
-                        headerGroup.headers.findIndex(
-                          (c) => c.id !== SELECTION_CELL_ID, // first non-selection column is the expander column
-                        );
-
-                    if ([undefined, 0].includes(column.minWidth)) {
-                      // override "undefined" or zero min-width with default value
-                      column.minWidth = columnHasExpanders
-                        ? COLUMN_MIN_WIDTHS.withExpander
-                        : COLUMN_MIN_WIDTHS.default;
-
-                      // set the minWidth to the user provided width instead if the width is less than the default minWidth
-                      if (
-                        typeof column.width === 'number' &&
-                        column.minWidth > column.width
-                      ) {
-                        column.minWidth = column.width;
-                      }
-                    }
-
-                    const columnProps = column.getHeaderProps({
-                      ...restSortProps,
-                      className: cx(
-                        'iui-table-cell',
-                        {
-                          'iui-actionable': column.canSort,
-                          'iui-sorted': column.isSorted,
-                          'iui-table-cell-sticky': !!column.sticky,
-                        },
-                        column.columnClassName,
-                      ),
-                      style: {
-                        ...getCellStyle(column, !!state.isTableResizing),
-                        ...(columnHasExpanders && getSubRowStyle({ density })),
-                        ...getStickyStyle(column, visibleColumns),
-                        flexWrap: 'wrap',
-                        columnGap: 'var(--iui-size-xs)',
-                      },
-                    });
-
+                    const dragAndDropProps = column.getDragAndDropProps();
                     return (
-                      <Box
-                        {...columnProps}
-                        {...column.getDragAndDropProps()}
-                        key={columnProps.key}
-                        title={undefined}
-                        ref={(el) => {
-                          columnResizeRef(el, column);
-                        }}
-                        onMouseDown={() => {
-                          isHeaderDirectClick.current = true;
-                        }}
-                        onClick={(e) => {
-                          // Prevents from triggering sort when resizing and mouse is released in the middle of header
-                          if (isHeaderDirectClick.current) {
-                            onClick?.(e);
-                            isHeaderDirectClick.current = false;
-                          }
-                        }}
-                        tabIndex={showSortButton(column) ? 0 : undefined}
-                        onKeyDown={(e) => {
-                          if (e.key == 'Enter' && showSortButton(column)) {
-                            column.toggleSortBy();
-                          }
-                        }}
-                      >
-                        <>
-                          <ShadowRoot>
-                            {typeof column.Header === 'string' ? (
-                              <LineClamp>
-                                <slot />
-                              </LineClamp>
-                            ) : (
-                              <slot />
-                            )}
-                            <slot name='actions' />
-                            <slot name='resizers' />
-                            <slot name='shadows' />
-                          </ShadowRoot>
-
-                          {column.render('Header')}
-                          {(showFilterButton(column) ||
-                            showSortButton(column)) && (
-                            <Box
-                              className='iui-table-header-actions-container'
-                              onKeyDown={(e) => e.stopPropagation()} // prevents from triggering sort
-                              slot='actions'
-                            >
-                              {showFilterButton(column) && (
-                                <FilterToggle column={column} />
-                              )}
-                              {showSortButton(column) && (
-                                <Box className='iui-table-cell-end-icon'>
-                                  {column.isSortedDesc ||
-                                  (!column.isSorted && column.sortDescFirst) ? (
-                                    <SvgSortDown
-                                      className='iui-table-sort'
-                                      aria-hidden
-                                    />
-                                  ) : (
-                                    <SvgSortUp
-                                      className='iui-table-sort'
-                                      aria-hidden
-                                    />
-                                  )}
-                                </Box>
-                              )}
-                            </Box>
-                          )}
-                          {isResizable &&
-                            column.isResizerVisible &&
-                            (index !== headerGroup.headers.length - 1 ||
-                              columnResizeMode === 'expand') && (
-                              <Box
-                                {...column.getResizerProps()}
-                                className='iui-table-resizer'
-                                slot='resizers'
-                              >
-                                <Box className='iui-table-resizer-bar' />
-                              </Box>
-                            )}
-                          {enableColumnReordering &&
-                            !column.disableReordering && (
-                              <Box
-                                className='iui-table-reorder-bar'
-                                slot='resizers'
-                              />
-                            )}
-                          {column.sticky === 'left' &&
-                            state.sticky.isScrolledToRight && (
-                              <Box
-                                className='iui-table-cell-shadow-right'
-                                slot='shadows'
-                              />
-                            )}
-                          {column.sticky === 'right' &&
-                            state.sticky.isScrolledToLeft && (
-                              <Box
-                                className='iui-table-cell-shadow-left'
-                                slot='shadows'
-                              />
-                            )}
-                        </>
-                      </Box>
+                      <ColumnHeader<T>
+                        {...dragAndDropProps}
+                        key={dragAndDropProps.key}
+                        columnRefs={columnRefs}
+                        column={column}
+                        index={index}
+                        areFiltersSet={areFiltersSet}
+                        hasAnySubRows={hasAnySubRows}
+                        headers={headerGroup.headers}
+                        state={state}
+                        data={data}
+                        isResizable={isResizable}
+                        columnResizeMode={columnResizeMode}
+                        enableColumnReordering={enableColumnReordering}
+                        density={density}
+                        visibleColumns={visibleColumns}
+                      />
                     );
                   })}
                 </Box>
